Separate TransactionListRow component into a reusable TransactionRow
